Updated planned contribution section
Re-written, somewhat better wording and more extensive description of
the actual contribution.

I also stopped syncing the other files, due to the .ignore part made all
the other files be hidden so I couldn't compile.

Used JabRef to auto-generate BibTex Keys, and updated @misc with
howpublished, since the URL doesn't show in unsrt bib-style
(http://tex.stackexchange.com/questions/186235/in-bibtex-when-should-i-use-howpublished-and-when-url)