Updated rev_history and bibtex file
Ran clean on all BibTex entries with JabRef. The original bibtex file is
now named .old, and the new is the one that has been cleaned and had all
cite-keys updated.
Since the research question in Sec 3.1 is more related to the result,
rather than existing research, I could not find anything relevant.
Therefore this chapter will not be included in the final pagecount (req:
3-10 pages, meaning at least 4+ pages must be written for this chapter
to be valid