Updates based on feedback
Updated based on feedback in #issue1.
see rev_history for changes.
Also managed to find a temp fix by adding natbib and changing bibstyle
when building to unsrtnat so that it also prints author names (it got
set back to default style after build, to avoid complaints in regards to
using other bibstyle)